requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
